Changes made on 14 july
- Implemented Google SSO Login
- Implemented email & PW signup/login & set up the DB for it
- Separated EntryDetailsModal, JournalEntryModal, SignaturePadModal, IDScannerModal
- Added scan-id button inside journal entry
