+++ package.json
@@ -10,9 +10,15 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@react-oauth/google": "0.12.2",
+    "@supabase/supabase-js": "^2.51.0",
+    "gapi-script": "^1.2.0",
+    "jwt-decode": "^4.0.0",
     "lucide-react": "^0.344.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.6.3",
+    "signature_pad": "^5.0.10"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
